+++ scripts/webpack.config.ts
@@ -106,8 +106,8 @@ export const generateConfig = (isProduction: boolean): webpack.Configuration => 
     resolve: {
       extensions: [".js", ".ts", ".tsx", ".scss", ".json"],
       alias: {
-        "@app/view-component": appPath("./src/view-component/index.ts"),
-        "@app/container-component": appPath("./src/container-component/index.ts"),
+        "@app/component": appPath("./src/component/index.ts"),
+        "@app/container": appPath("./src/container/index.ts"),
         "@app/domain": appPath("./src/domain/index.ts"),
         "@app/infra": appPath("./src/infra/index.ts"),
       },
