+++ src/App.tsx
@@ -1,6 +1,6 @@
 import * as Domain from "@app/domain";
 import * as React from "react";
-import { Page } from "./container-component";
+import { Page } from "./container";
 import "./style.scss";
 
 interface AppProps {
