+++ src/component/Counter/counter.scss
