+++ src/component/Page/page.scss
