+++ src/component/index.ts
